small change to utils_mod
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@493 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/utils_mod.f90
@@ -2017,7 +2017,7 @@ contains
   subroutine convertByte2(iByte, ival)
     integer(kind=single) :: ibyte(2), iswap
     integer(kind=single) :: ubyte(2)
-    integer(kind=single)  :: ival
+    integer(kind=double)  :: ival
 
     ubyte = ibyte
     where (ubyte < 0) ubyte = ibyte + 256
